% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wzyadd.R
\name{wzyadd}
\alias{wzyadd}
\title{Add together two numbers.}
\usage{
wzyadd(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function add two numbers together, which written by wzy.
}
\examples{
add(1, 1)
add(10, 1)
}
